class split
