+++ app/src/main/java/ru/practicum/android/diploma/util/App.kt
@@ -21,4 +21,10 @@ class App : Application() {
             )
         }
     }
+    companion object {
+        const val REGION_ID_KEY = "REGION_ID_KEY"
+        const val PLACE_OF_WORK_KEY = "PLACE_OF_WORK_KEY"
+        const val PLACE_OF_WORK_COUNTRY_KEY = "PLACE_OF_WORK_COUNTRY_KEY"
+        const val PLACE_OF_WORK_REGION_KEY = "PLACE_OF_WORK_REGION_KEY"
+    }
 }
